исправления scale
